[es6-bookreader-core/debugConsole] create debugConsole.js
